backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlmodel==0.0.37
SQLAlchemy==2.0.48
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-jose==3.5.0
psycopg[binary]==3.3.3
alembic==1.16.5
pwdlib==0.3.0